+++ sdk/azidentity/CHANGELOG.md
@@ -4,6 +4,15 @@
 ### Breaking Changes
 * The `AuthorityHost` field in credential options structs is now a custom type,
   `AuthorityHost`, with underlying type `string`
+* `NewChainedTokenCredential` has a new signature to accommodate a placeholder
+  options struct:
+  ```go
+  // before
+  cred, err := NewChainedTokenCredential(credA, credB)
+
+  // after
+  cred, err := NewChainedTokenCredential([]azcore.TokenCredential{credA, credB}, nil)
+  ```
 
 
 ## 0.11.0 (2021-09-08)